% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evaluate}
\alias{evaluate}
\title{Do evaluation.}
\usage{
evaluate(pop, model_data, model = "lm", glm_family = NULL,
  criterion = "AIC", do_parallel = FALSE)
}
\arguments{
\item{pop}{Matrix of population of chromosomes.}

\item{model_data}{Object of class model_data.}

\item{model}{Character; "lm" (default) or "glm"; Linear model or
generalized linear model.}

\item{glm_family}{Character if model is "glm", NULL otherwise;
"binomial", "gaussian" (default), "Gamma", "inverse.gaussian", "poisson", "quasi",
"quasibinomial", "quasipoisson"; A family function that gives the error
distribution and link function to be used in the model.}

\item{criterion}{"AIC" (default) or "BIC"; Criterion to be minimized.}

\item{do_parallel}{Logical; Default FALSE; Do in parallel?}
}
\value{
Numeric vector; Evaluation values for all chromosomes
in the current generation.
}
\description{
Do evaluation for chromosomes in population by calculating model selection criterion.
}

