% Generated by roxygen2 (4.0.2): do not edit by hand
\name{evaluate_once}
\alias{evaluate_once}
\title{Do evaluation once.}
\usage{
evaluate_once(model_data, xvars_select, model = "lm", glm_family = NULL,
  criterion = "AIC")
}
\arguments{
\item{model}{Character; "lm" (default) or "glm"; Linear model or generalized linear model.}

\item{glm_family}{Character if model is "glm", NULL otherwise;
"binomial", "gaussian" (default), "Gamma", "inverse.gaussian", "poisson", "quasi",
"quasibinomial", "quasipoisson"; A family function that gives the error
distribution and link function to be used in the model.}

\item{criterion}{"AIC" (default) or "BIC"; AIC or BIC.}

\item{model_data;}{Object of class model_data.}

\item{xvars_select;}{Logical vector;}
}
\value{
Numeric; Value of criterion.
}
\description{
Do evaluation for a chromosome by calculating model selection criterion.
}

