% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.ga}
\alias{plot.ga}
\title{Plots results from the genetic algorithim.}
\usage{
\method{plot}{ga}(ga, num_view = 3)
}
\arguments{
\item{ga}{Object of class ga.}

\item{num_view}{Number of top models to display.}
}
\value{
Prints summary of top models and associated value of
model selection criterion.
}
\description{
Plots the best model evaluation criterion in each generation
against the generation iteration.
}

