% Generated by roxygen2 (4.0.2): do not edit by hand
\name{process_data}
\alias{process_data}
\title{Process data for input into genetic algorithm.}
\usage{
process_data(data, yvar, xvars = NULL)
}
\arguments{
\item{data}{Data frame}

\item{yvar}{Character; Name of column containing response variable.}

\item{xvars}{Character vector; Default is all column names that are not yvar;
Name(s) of column(s) containing set of explanatory variables to select on.}
}
\value{
A list object named model_data containing:
\describe{
  \item{data}{Data frame; Processed data with only relevant columns.}
  \item{yvar}{Character; Name of column containing response variable.}
  \item{xvars}{Character vector; Name(s) of column(s) containing
  set of explanatory variables to select on.}
  \item{num_vars}{Integer; Length of xvars.}
}
}
\description{
Process data for input into genetic algorithm.
}

