% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recombine}
\alias{recombine}
\title{Recombine.}
\usage{
recombine(pop_mating, pop_size, method = "onepoint", prob_recombine = 0.6,
  do_parallel = FALSE)
}
\arguments{
\item{pop_mating}{Matrix of population of chromosomes that form the mating pool.}

\item{pop_size}{Integer; Number of chromosomes in a generation.}

\item{method}{String; "onepoint" (default), "twopoint", "uniform";
Type of crossover, at one point, at two points or uniformly (at all possible points).}

\item{prob_recombine}{Numeric, between 0 and 1; Default is 0.6;
Probability of recombination.}

\item{do_parallel}{Logical; Default FALSE; Do in parallel?}
}
\value{
Matrix of population of chromosomes resulting from recombination.
}
\description{
Carry out crossover of parent chromosomes in a mating pool.
}

