% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reproduce}
\alias{reproduce}
\title{Wrapper function for reproduction stage.}
\usage{
reproduce(ga, iteration, do_parallel = FALSE)
}
\arguments{
\item{ga}{Object of class ga.}

\item{iteration}{Iteration number.}
}
\value{
Updated ga list object.
}
\description{
Wrapper function for reproduction stage.
}

