% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.ga}
\alias{summary.ga}
\title{Display summary of results from the genetic algorithim.}
\usage{
\method{summary}{ga}(ga, num_view = 5)
}
\arguments{
\item{ga}{Object of class ga.}

\item{num_view}{Number of top models to display.}
}
\value{
Prints summary of top models and associated value of
model selection criterion.
}
\description{
Outputs the top models selected from the genetic algorithm.
}

